initial transition to react-google-maps
