+++ react-front-end/package.json
@@ -17,6 +17,7 @@
     "react": "^17.0.2",
     "react-animated-css": "^1.2.1",
     "react-dom": "^17.0.2",
+    "react-google-maps": "^9.4.5",
     "react-hover-image": "^2.0.1",
     "react-scripts": "4.0.3",
     "react-wordcloud": "^1.2.7",
